flip logic back around

+++ libraries/render-utils/src/Model.h
@@ -389,7 +389,7 @@ protected:
 
     QUrl _url;
     bool _isVisible;
-    uint32_t _viewVisibilityMask { 0 };
+    uint8_t _viewVisibilityMask { 0 };
 
     gpu::Buffers _blendedVertexBuffers;
